pyproject.toml
[tool.poetry]
name = "social"
version = "0.1.0"
description = ""
authors = ["rbc33 <ricardobenthem@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.8.2"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
jinja2 = "^3.1.4"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
pyjwt = "^2.9.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
